serverからactionを切り離した
